fix/refactor(levelstats): move to utils, check locked/hidden, doAll prod ready
